third commit

+++ src/Main.java
@@ -6,7 +6,7 @@ public class Main {
 
         int water = 3000; // миллилитров воды
         int potatoes = 5; // картофелин
-        double chicken = 6; // куриных бёдер
+        int chicken = 6; // куриных бёдер
         int spices = 10; // ложек специй
 
         System.out.println("Сварили суп. На одного человека вышло:");
